refactor(core,schemas): refactor the hook library code
refactor the webhooks library code. address some comments
