+++ .changeset/metal-lions-swim.md
@@ -2,4 +2,7 @@
 "@logto/shared": patch
 ---
 
-add normalizeError method to @logto/shared package. Use this method to normalize error objects for logging. This method is useful for logging errors in a consistent format.
+add `normalizeError` method to `@logto/shared` package.
+
+Use this method to normalize error objects for logging.
+This method is useful for logging errors in a consistent format.
